+++ CHANGELOG.md
@@ -2,6 +2,49 @@
 
 All notable changes to this project will be documented in this file. This project uses [Semantic Versioning](https://semver.org/).
 
+## [Beta Version 5.0.0](https://github.com/donavanbecker/homebridge-meross/compare/v4.0.0....5.0.0) (2021-XX-XX)
+
+### Breaking Changes
+
+- The Plugin has been been changed from an `accessory` type to a `platform` type.
+- You will have to change you config completely if you update to this version.
+  - You can take your current `accessory` and move it to the platform config.
+  - Example: 
+    - `Before`:
+    ```json
+    {
+      "model": "MSS620",
+      "name": "Outlet",
+      "deviceUrl": "192.168.1.1",
+      "channel": 0,
+      "messageId": "abcdefghijklmnopqrstuvwxyz123456789",
+      "timestamp": 123456789,
+      "sign": "abcdefghijklmnopqrstuvwxyz123456789",
+      "accessory": "Meross"
+    }
+    ```
+    - `After`:
+    ```json
+    {
+    "name": "Meross",
+    "devices": [
+        // This is the from above
+        {
+        "model": "MSS620",
+        "name": "Outlet",
+        "deviceUrl": "192.168.1.1",
+        "channel": 0,
+        "messageId": "abcdefghijklmnopqrstuvwxyz123456789",
+        "timestamp": 123456789,
+        "sign": "abcdefghijklmnopqrstuvwxyz123456789",
+        "accessory": "Meross" //Remove this.
+        }
+        // Ends here from above
+      ],
+    "platform": "Meross"
+    }
+    ```
+
 ## [Version 4.0.0](https://github.com/donavanbecker/homebridge-meross/compare/v3.5.0....4.0.0) (2021-3-02)
 
 ### Major Changes